I think this is a simpler patch and reuses more code with original.
